Updated commons-fs2 for cats effect 3

+++ modules/common-fs2/src/main/scala/com/snowplowanalytics/snowplow/enrich/common/fs2/config/io.scala
@@ -330,7 +330,7 @@ object io {
       deriveConfiguredEncoder[Output]
   }
 
-  final case class Concurrency(enrich: Long, sink: Int)
+  final case class Concurrency(enrich: Int, sink: Int)
 
   object Concurrency {
     implicit val concurrencyDecoder: Decoder[Concurrency] =
